move computeSquareCenter from darg.ts to util.ts in order to eliminate cyclic depenedency
